Separate out ModuleDepGraphNodes
